Merfe JsonEscaper class into JsonTypeSerializer.Unescape
